# docs/swagger.yaml
basePath: /
definitions:
  dto.AdminCreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  dto.AdminUpdateUserActiveRequest:
    properties:
      is_active:
        type: boolean
    type: object
  dto.AdminUpdateUserRoleRequest:
    properties:
      role:
        type: string
    type: object
  dto.AdminUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      role:
        type: string
    type: object
  dto.AdminUsersListResponse:
    properties:
      count:
        type: integer
      status:
        type: string
      users:
        items:
          $ref: '#/definitions/dto.AdminUserResponse'
        type: array
    type: object
...
  dto.AuthTokenResponse:
    properties:
      access_token:
        type: string
      email:
        type: string
      expires_at:
        type: string
      expires_in:
        type: integer
      refresh_expires_at:
        type: string
      refresh_expires_in:
        type: integer
      refresh_token:
        type: string
      role:
        type: string
      token_type:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.CreatedUserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  dto.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  dto.RegisterStudentRequest:
    properties:
      course:
        type: integer
      email:
        type: string
      fullname:
        type: string
      group:
        type: string
      password:
        type: string
      uid:
        type: string
    type: object
  dto.RevokeRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  http.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
      message:
        example: course query parameter is required
        type: string
      status:
        example: error
        type: string
    type: object
...
info:
  contact: {}
  description: API for schedule import/read, attendance processing and study plan
    management.
  title: CourseJob API
  version: "1.0"
paths:
  /admin/users:
    get:
      description: Returns users with optional role/email filters. Requires Bearer
        token with admin role.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Role filter: student|teacher|admin'
        in: query
        name: role
        type: string
      - description: Email contains filter
        in: query
        name: email
        type: string
      - description: Limit (default 100, max 200)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AdminUsersListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: List users by admin
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Creates teacher or admin account. Requires Bearer token with admin
        role.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Create user payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreatedUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Create user by admin
      tags:
      - Admin
  /admin/users/{id}/active:
    patch:
      consumes:
      - application/json
      description: Updates user is_active status. Requires Bearer token with admin
        role.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Active update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpdateUserActiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AdminUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Update user active flag
      tags:
      - Admin
  /admin/users/{id}/role:
    patch:
      consumes:
      - application/json
      description: 'Updates user role. Allowed roles: teacher or admin. Requires Bearer
        token with admin role.'
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpdateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AdminUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Update user role
      tags:
      - Admin
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates by email and password and returns access token.
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthTokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Issues a new access/refresh token pair using a valid refresh token.
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Refresh token
      tags:
      - Auth
  /auth/register/student:
    post:
      consumes:
      - application/json
      description: Registers a student user and creates both student profile and user
        account in one transaction.
      parameters:
      - description: Student registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Register student
      tags:
      - Auth
  /auth/revoke:
    post:
      consumes:
      - application/json
      description: Revokes a refresh token so it can no longer be used.
      parameters:
      - description: Revoke token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RevokeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Revoke token
      tags:
      - Auth
...
schemes:
- https
swagger: "2.0"